refactor(js): split quickjs bindings
